refactor(tui): fold comment pagination into screens/detail.rs
The comment-pagination block (`reset_comment_pagination`,
`apply_initial_comments`, `apply_older_comments`,
`can_load_older_comments`, `InitialComments`) only ever touches
`DetailState`'s own fields, even though its callers live in
`dispatch.rs`/`bg.rs`. `screens/detail.rs` is already the established
home for GistDetail-scoped logic (#287, #310), so this folds the
block in there instead of splitting one screen's logic across a new
sibling file. `InitialComments` re-exported flat from `mod.rs` so
`crate::tui::InitialComments` keeps resolving unchanged at its
existing call sites. No behavior change.

Part of #313
